update Andrew 04-26-2015
I have had a little bit of time to work on the game. Here is what i have
done.

-we now have a movable player
-we now have our own algorithm for the players gravity, acceleration,
and friction
-i made moving background clouds
-i made a simple background song
-an interactive npc named Mr. Flower has been added to the game (this is
just to demonstrate npc dialog Mr. Flower probably wont be in the final
game)
-sand texture has been added
-camera scrolls with the player

I have attracted a standalone copy of the current version of the game. I
have also added it to the repository so you will be able to work with it
when you get the tools you need. I will probably be able to give them to
you next week. See if you like the game, its ok if you don't, it didn't
take me long to create it and it is still in very early development. Its
playable but not fun yet. Below are some things we need to discuss and
think about.

-will our game have levels?
-is this a puzzle game or one that has a focus on skill
-is this an open world game?
-enemys, what type of creatures will our player fight?
-you mentioned something about some type of leveling up system, however,
we need to plan this out in more detail
-is huger or thirst a problem for our character
-can the player buy weapons?

So far I've been kind of going for an apocalyptic setting. I will
eventually replace our player with something other than a square, I just
haven't got around to it. A second player can be added to the game
easily. I haven't made a main menu to the game but i will do that
eventually. Let me know what you think about it so far. Try to start
planing things out.